requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain-groq
langchain-huggingface
langchain-text-splitters
langchain-community
langchain-core
langchain-pinecone
pinecone-client
pypdf
sentence-transformers
jinja2